[VIEW-ONLY] dtype climb + WS1 fortify (reduction seed) — PR-bound code, squashed for review
Synthetic single-commit diff of PRODUCT CODE only (examples/ + helion/);
_lab/ reasoning artifacts (6877 lines) intentionally excluded. NOT real history; do not merge.

Squashed range: baseline e9251140 -> WS1 tip 4b65f2ef (dtype climb + the 22 [ws1] commits),
restricted to non-_lab files (13 files, +844/-23).

  examples/{jsd,sum,welford,kl_div}.py        dtype correctness (fp32-accumulate, jsd accumulator, fp16 notes)
  examples/{groupnorm,log_softmax,logsumexp,row_max,l2_norm,argmax}.py  NEW WS1 fortify siblings + coverage probes
  helion/_compiler/autotuner_heuristics/triton.py   _num_warps bands (dtype w8/w1) + WS1 levers
  helion/_compiler/device_ir.py                      fact populators
  helion/autotuner/config_spec.py                    ReductionFact fields

+++ examples/welford.py
@@ -48,7 +48,13 @@ def welford(
         acc_m2 = torch.zeros_like(acc_cnt)
 
         for tile_n in hl.tile(n):
-            chunk = x[tile_m, tile_n]
+            # Upcast the loaded chunk to fp32 before the sum / sum-of-squares. At
+            # bf16/fp16 an input-width accumulator loses precision badly, and in fp16
+            # sum_x*sum_x (a per-tile partial ~ (W/2)^2) overflows 65504 -> negative
+            # variance -> rsqrt(neg) -> NaN. Accumulating in fp32 (what torch
+            # layer_norm does) fixes both. The HBM load is still x.dtype, so no extra
+            # bandwidth; no-op at fp32 input.
+            chunk = x[tile_m, tile_n].to(torch.float32)
             # Count of VALID columns. OOB loads are masked to 0, so sum_x/sum_x2 are
             # already correct; only the divisor must use the true valid count, NOT the
             # constexpr tile width chunk.size(-1) (over-counts last-tile padding at
